Add External Helpers and Bypass since we need HttpApp
